update minimal script to converge towards rails new's defaults

+++ minimal.rb
@@ -2,38 +2,21 @@ run 'pgrep spring | xargs kill -9'
 
 # GEMFILE
 ########################################
-run 'rm Gemfile'
-file 'Gemfile', <<-RUBY
-source 'https://rubygems.org'
-ruby '#{RUBY_VERSION}'
-
-#{"gem 'bootsnap', require: false" if Rails.version >= "5.2"}
-gem 'jbuilder', '~> 2.0'
-gem 'pg', '~> 0.21'
-gem 'puma'
-gem 'rails', '#{Rails.version}'
-gem 'redis'
-
-gem 'autoprefixer-rails'
-gem 'font-awesome-sass', '~> 5.6.1'
-gem 'sassc-rails'
-gem 'simple_form'
-gem 'uglifier'
-gem 'webpacker'
-
-group :development do
-  gem 'web-console', '>= 3.3.0'
+inject_into_file 'Gemfile', before: 'group :development, :test do' do
+  <<-RUBY
+    gem 'autoprefixer-rails'
+    gem 'font-awesome-sass'
+    gem 'simple_form'
+  RUBY
 end
 
-group :development, :test do
-  gem 'pry-byebug'
-  gem 'pry-rails'
-  gem 'listen', '~> 3.0.5'
-  gem 'spring'
-  gem 'spring-watcher-listen', '~> 2.0.0'
-  gem 'dotenv-rails'
+inject_into_file 'Gemfile', after: 'group :development, :test do' do
+  <<-RUBY
+    gem 'pry-byebug'
+    gem 'pry-rails'
+    gem 'dotenv-rails'
+  RUBY
 end
-RUBY
 
 # Ruby version
 ########################################
@@ -42,7 +25,7 @@ file '.ruby-version', RUBY_VERSION
 # Procfile
 ########################################
 file 'Procfile', <<-YAML
-web: bundle exec puma -C config/puma.rb
+  web: bundle exec puma -C config/puma.rb
 YAML
 
 # Assets
@@ -52,11 +35,13 @@ run 'rm -rf vendor'
 run 'curl -L https://github.com/lewagon/stylesheets/archive/master.zip > stylesheets.zip'
 run 'unzip stylesheets.zip -d app/assets && rm stylesheets.zip && mv app/assets/rails-stylesheets-master app/assets/stylesheets'
 
-run 'rm app/assets/javascripts/application.js'
-file 'app/assets/javascripts/application.js', <<-JS
-//= require rails-ujs
-//= require_tree .
-JS
+if Rails.version < "6"
+  run 'rm app/assets/javascripts/application.js'
+  file 'app/assets/javascripts/application.js', <<-JS
+    //= require rails-ujs
+    //= require_tree .
+  JS
+end
 
 # Dev environment
 ########################################
@@ -64,42 +49,35 @@ gsub_file('config/environments/development.rb', /config\.assets\.debug.*/, 'conf
 
 # Layout
 ########################################
-run 'rm app/views/layouts/application.html.erb'
-file 'app/views/layouts/application.html.erb', <<-HTML
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <title>TODO</title>
-    <%= csrf_meta_tags %>
-    <%= action_cable_meta_tag %>
-    <%= stylesheet_link_tag 'application', media: 'all' %>
-    <%#= stylesheet_pack_tag 'application', media: 'all' %> <!-- Uncomment if you import CSS in app/javascript/packs/application.js -->
-  </head>
-  <body>
-    <%= yield %>
-    <%= javascript_include_tag 'application' %>
-    <%= javascript_pack_tag 'application' %>
-  </body>
-</html>
-HTML
+if Rails.version < "6"
+  scripts = <<-HTML
+    <%= javascript_include_tag 'application', 'data-turbolinks-track': 'reload', defer: true %>"
+    <%= javascript_pack_tag 'application', 'data-turbolinks-track': 'reload' %>
+  HTML
+  gsub_file('app/views/layouts/application.html.erb', "<%= javascript_include_tag 'application', 'data-turbolinks-track': 'reload' %>", scripts)
+end
+gsub_file('app/views/layouts/application.html.erb', "<%= javascript_pack_tag 'application', 'data-turbolinks-track': 'reload' %>", "<%= javascript_pack_tag 'application', 'data-turbolinks-track': 'reload', defer: true %>")
+inject_into_file 'app/views/layouts/application.html.erb', after: "<%= stylesheet_link_tag 'application', media: 'all' %>" do
+  <<-HTML
+    <%#= stylesheet_pack_tag 'application', media: 'all' %>
+  HTML
+end
 
 # README
 ########################################
 markdown_file_content = <<-MARKDOWN
-Rails app generated with [lewagon/rails-templates](https://github.com/lewagon/rails-templates), created by the [Le Wagon coding bootcamp](https://www.lewagon.com) team.
+  Rails app generated with [lewagon/rails-templates](https://github.com/lewagon/rails-templates), created by the [Le Wagon coding bootcamp](https://www.lewagon.com) team.
 MARKDOWN
 file 'README.md', markdown_file_content, force: true
 
 # Generators
 ########################################
 generators = <<-RUBY
-config.generators do |generate|
-      generate.assets false
-      generate.helper false
-      generate.test_framework  :test_unit, fixture: false
-    end
+  config.generators do |generate|
+    generate.assets false
+    generate.helper false
+    generate.test_framework  :test_unit, fixture: false
+  end
 RUBY
 
 environment generators
@@ -122,39 +100,37 @@ after_bundle do
   ########################################
   append_file '.gitignore', <<-TXT
 
-# Ignore .env file containing credentials.
-.env*
+    # Ignore .env file containing credentials.
+    .env*
 
-# Ignore Mac and Linux file system files
-*.swp
-.DS_Store
-TXT
+    # Ignore Mac and Linux file system files
+    *.swp
+    .DS_Store
+  TXT
 
   # Webpacker / Yarn
   ########################################
-  run 'rm app/javascript/packs/application.js'
   run 'yarn add popper.js jquery bootstrap'
-  file 'app/javascript/packs/application.js', <<-JS
-import "bootstrap";
-JS
+  file_prepend 'app/javascript/packs/application.js', <<-JS
+    import "bootstrap";
+  JS
 
   inject_into_file 'config/webpack/environment.js', before: 'module.exports' do
-<<-JS
-const webpack = require('webpack')
+    <<-JS
+      const webpack = require('webpack');
 
-// Preventing Babel from transpiling NodeModules packages
-environment.loaders.delete('nodeModules');
+      // Preventing Babel from transpiling NodeModules packages
+      environment.loaders.delete('nodeModules');
 
-// Bootstrap 4 has a dependency over jQuery & Popper.js:
-environment.plugins.prepend('Provide',
-  new webpack.ProvidePlugin({
-    $: 'jquery',
-    jQuery: 'jquery',
-    Popper: ['popper.js', 'default']
-  })
-)
-
-JS
+      // Bootstrap 4 has a dependency over jQuery & Popper.js:
+      environment.plugins.prepend('Provide',
+        new webpack.ProvidePlugin({
+          $: 'jquery',
+          jQuery: 'jquery',
+          Popper: ['popper.js', 'default']
+        })
+      );
+    JS
   end
 
   # Dotenv
@@ -167,7 +143,6 @@ JS
 
   # Git
   ########################################
-  git :init
   git add: '.'
   git commit: "-m 'Initial commit with minimal template from https://github.com/lewagon/rails-templates'"
 end
